fix SQL keyword document
